requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
SQLAlchemy==2.0.21
pydantic==2.3.0
pymysql==1.1.0
cryptography==41.0.3  # Required for PyMySQL to connect securely
python-dotenv==1.0.0
docker>=5.0.0
